requirements.txt
pytest~=6.2.3
Django~=3.2.8
djangorestframework~=3.12.4
requests~=2.25.1